BCH - TLL PR feedback
