+++ templates/admin/assets/scss/_root.scss
@@ -18,7 +18,7 @@ button {
 
 hr {
   border: 0;
-  border-top: 1px solid $separator-color;
+  border-top: $separator;
 }
 
 /* Remove default margin */
